requirements.txt
Django==3.2.5
djangorestframework==3.13.1
django-cors-headers==3.13.0
mysqlclient==2.1.0
requests==2.28.1
beautifulsoup4==4.11.1